Add sensors and debug parameters to config file
The list 'sensors' contains the list of sensors which should be checked
and sent to mqtt broker.

'Debug' parameter controls if data should be printed to output or sent
to mqtt broker.

It allows to change the configuration without modifying code.

Also move the 1st definition of sensors outside the main cycle - no need
to do it constantly.